import-maps/csp/unsafe-inline.tentative.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';">
<script type="importmap">
{
  "imports": {
    "../resources/log.sub.js?name=A": "../resources/log.sub.js?name=B"
  }
}
</script>
<script>
const log = [];
promise_test(() => {
  return import("../resources/log.sub.js?name=A")
    .then(() => assert_array_equals(log, ["log:B"]))
  },
  'Importmap should be accepted due to unsafe-inline');
</script>
